Move util dir to utils to avoid name conflict.
