+++ __tests__/serverjs/updatecards.test.js
@@ -4,7 +4,7 @@ const rimraf = require('rimraf');
 const updatecards = require('../../serverjs/updatecards');
 const carddb = require('../../serverjs/cards');
 const examplecards = require('../../fixtures/examplecards');
-const cardutil = require('../../dist/util/Card.js');
+const cardutil = require('../../dist/utils/Card.js');
 
 const cardsFixturePath = 'fixtures/cards_small.json';
