+++ __tests__/src/filter.test.js
@@ -2,7 +2,7 @@ const carddb = require('../../serverjs/cards');
 const fixturesPath = 'fixtures';
 const cubefixture = require('../../fixtures/examplecube');
 
-import Filter from '../../src/util/Filter';
+import Filter from '../../src/utils/Filter';
 import { expectOperator } from '../helpers';
 
 const setCounts = (cards, propertyName) => {
