+++ src/utils/Affiliate.js
