+++ src/utils/Hash.js
